requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
aiohttp==3.8.6
bs4==0.0.1
beautifulsoup4==4.12.2
pandas==2.1.1
openpyxl==3.1.2
Pillow==10.1.0
python-multipart==0.0.6
aiofiles==23.2.1
databases==0.8.0
aiosqlite==0.19.0
SQLAlchemy==2.0.22